+++ arangod/Aql/Optimizer/Rules/CMakeLists.txt
@@ -1,4 +1,3 @@
 target_sources(arango_aql PRIVATE
     EnumeratePathsFilterMatcher.cpp
-    EnumeratePathsRule.cpp
-    ExpressionMatcher.cpp)
+    EnumeratePathsRule.cpp)
