+++ arangod/Aql/Optimizer/Rules/EnumeratePathsRule.cpp
@@ -56,6 +56,7 @@ void optimizeEnumeratePathsRule(Optimizer* opt,
     EnumeratePathsFilterMatcher f(plan.get());
     leaf->walk(f);
   }
+  ADB_PROD_ASSERT(false);
 
   opt->addPlan(std::move(plan), rule, appliedAChange);
 }
